+++ atlas.com/buddies/list/model.go
@@ -9,7 +9,7 @@ type Model struct {
 	tenantId    uuid.UUID
 	id          uuid.UUID
 	characterId uint32
-	capacity    uint32
+	capacity    byte
 	buddies     []buddy.Model
 }
 
@@ -17,6 +17,6 @@ func (m Model) Buddies() []buddy.Model {
 	return m.buddies
 }
 
-func (m Model) Capacity() uint32 {
+func (m Model) Capacity() byte {
 	return m.capacity
 }
